move `type_check` out of transform and into `type_check` module
